+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 IMPROVEMENT
 
 - [p2p] Add seed nodes
+- [sync] Speed up synchronization
 
 ## 0.2.4
 *Aug 24th, 2018*