+++ CHANGELOG.md
@@ -13,6 +13,7 @@ IMPROVEMENT
 - [sync] Speed up synchronization
 - [core] Extend max payload size to 1024 bytes
 - [api] Limit balance watchers to 10 clients
+- [config] Add config file
 
 ## 0.2.4
 *Aug 24th, 2018*